requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
dateparser==1.2.0
requests==2.32.3
pydantic==2.9.2
